Move container-only code into modal.runtime
