+++ modal/runtime/execution_context.py
@@ -2,9 +2,10 @@
 from contextvars import ContextVar
 from typing import Callable, List, Optional
 
-from modal._container_io_manager import _ContainerIOManager
 from modal._utils.async_utils import synchronize_api
 from modal.exception import InvalidError
+
+from ._container_io_manager import _ContainerIOManager
 
 
 def is_local() -> bool:
