+++ test/conftest.py
@@ -34,7 +34,6 @@ from grpclib.events import RecvRequest, listen
 
 import modal._serialization
 from modal import __version__, config
-from modal._container_io_manager import _ContainerIOManager
 from modal._serialization import serialize_data_format
 from modal._utils.async_utils import asyncify, synchronize_api
 from modal._utils.grpc_testing import patch_mock_servicer
@@ -45,6 +44,7 @@ from modal.app import _App
 from modal.client import Client
 from modal.image import ImageBuilderVersion
 from modal.mount import client_mount_name
+from modal.runtime._container_io_manager import _ContainerIOManager
 from modal_proto import api_grpc, api_pb2
