+++ test/webhook_test.py
@@ -7,10 +7,10 @@ import sys
 from fastapi.testclient import TestClient
 
 from modal import App, asgi_app, web_endpoint, wsgi_app
-from modal._asgi import webhook_asgi_app
 from modal.exception import DeprecationError, InvalidError
 from modal.functions import Function
 from modal.running_app import RunningApp
+from modal.runtime._asgi import webhook_asgi_app
 from modal_proto import api_pb2
 
 app = App()
